Implement order approval and payment link generation
This commit adds an order approval feature including the creation of an approval email with a respective payment link. The emailing system is handled by a new Mailable (OrderApproved) which notifies users about their order approval and provides a payment link if necessary. The commit also modifies the Order model to load 'orderItems'. Moreover, adjustments are made to create a CheckoutRequest with the Adyen Payment API to generate a payment link. Several other small changes are also included, primarily relating to event handling.