Disable SQL primary key requirement in database migration
The commit adds a new line of code to the database migration file for the users table. It uses the `DB::statement` method from the Laravel Illuminate Support package to set the `sql_require_primary_key` session variable to 0, effectively disabling the SQL requirement for primary keys on tables. This change can help to avoid issues when running migrations in certain database environments that enforce this requirement.